data: Make moving the offset implicitly increase the size.
Previously something like:

  nbdkit data data='@512'

would create a zero-sized disk.  This is because the size of the disk
is only increased by writing (even writing zeroes).

However this is both unexpected and it makes it awkward to use the new
@^ALIGNMENT operator to round up the disk size, eg to a whole number
of sectors.  The obvious does not work:

  nbdkit data data='<file @^512'

This change makes moving the offset beyond the end of the current disk
implicitly increase the size of the disk, so the above now works.  It
also works for nested patterns, eg:

  nbdkit data data='( "Hello" @^8 )*2'

cleanly rounds up the subpattern to 8 bytes.

An alternative would be adding another operator to do this, eg. "@."
but that seems unintuitive.

This also adds a test for this feature which doubles as a regression
test for the previous commit.