server: Permit - in plugin names
Use of - does not need shell quoting, and aids legibility in
multi-word plugin or filter names; it was already permitted by
is_short_name() in wrapper.c and main.c, only to be rejected later
during .load.  Permitting both - and _ would be ambiguous (not to
mention that things like 'man nbdkit-foo_bar-plugin' would look ugly),
so prefer only the character that is easier for human use.  Permitting
a leading - would be ambiguous with options, but restricting to a
letter as the first character would rule out any existing plugins such
as a theoretical '9p' plugin for interaction with the 9p filesystem.

Signed-off-by: Eric Blake <eblake@redhat.com>