sh: Fix off-by-one in INTERLEAVED .list_exports
We were happily searching beyond the end of s if the user supplied an
odd number of lines after the INTERLEAVED header.  This showed up as
intermittent failures in test-eval-exports.sh showing more names than
it was supposed to.  Also, make the test more robust to a current
working directory containing spaces.

Fixes: a2f199a091
Signed-off-by: Eric Blake <eblake@redhat.com>