New ondemand plugin.
This creates filesystems on demand.  A client simply connects with a
desired export name and a new export is created.  The export is
persistent (until deleted by the server admin), and clients may
disconnect and reconnect.  In some respects this is similar to
nbdkit-tmpdisk-plugin, or nbdkit-file-plugin with the dir= option.

Thanks: Eric Blake for both suggesting the idea behind the plugin and
providing feedback on the implementation.