server/log-syslog.c: Skip open_memstream on platforms which don't have it.
We only use open_memstream in a couple of places in the server, and
only in one place where it could be used by the Windows port.  In this
instance it's easy to avoid it, so do that.

Writing an open_memstream implementation for Windows is painful.  It
seems the best approach is to open a temporary file, write there, and
then read it back.  There is an MIT-licensed library of code here
which might be useful in future: https://github.com/Snaipe/fmem