build: Don't consider Cygwin to be the Windows platform.
We decided to take the approach of porting to the Windows native API
rather than using a Unix compatibility library like Cygwin.  This
means we only consider MinGW as the Windows platform.  In the somewhat
unlikely event that we wanted to support Cygwin in future, we would
consider it to be a completely new platform and treat it differently
from what we're currently calling "Windows".

Note that MSYS2 (not MSYS) is a lightweight Cygwin.  I don't know what
that looks like, but if we need to support it in future then it would
be grouped with Cygwin.

Thanks: Eric Blake

Updates: commit 1dd86376dc3933f657b25ad01f59a6a8e228f5d3