build: On Windows every plugin/filter must link with the import library.
In the previous commit we built an import library on Windows
(server/libnbdkit.a).  This contains the stubs which allow plugins to
dynamically link to symbols such as nbdkit_error exported by
nbdkit.exe.

An additional complexity is added by libtool which refuses to believe
that it's possible to add a *.a file when generating a dynamic
library.  We must therefore hide the *.a file from libtool by using
-Wl options (ie. passing the library directly through to the linker).