tests: Add some autoconf-defined variables to functions.sh.
Previously we passed variables @CXX@, @OCAMLOPT@ and @PYTHON@ through
TESTS_ENVIRONMENT.  Since these are only used in test shell scripts
that ‘source functions.sh’ and since that file is generated by
configure, we may as well add these variables to functions.sh, and
it's a little bit cleaner too.

This is refactoring and should make no difference to the tests.