data, memory: Add mlock=(yes|no) to --dump-plugin output.
Since allocator=malloc,mlock=true was made conditional on the platform
having support for mlock(2), it makes sense to allow users to discover
if the feature is supported.  This also adjusts the one test of this
feature to do this discovery before trying to test it.

Updates: commit 6c7e7d1f0ba9ad6bff927bc7922c970cfd5d249f