tls-fallback: Avoid next_open when !is_tls
As mentioned in TODO, the only reason we had to call next_open was
because of an assertion in backend.c.  It turns out fixing the server
to not care if a plugin calls next_open was fairly simple, letting us
fix that flaw of non-deterministic time spent in next_open even when
we have nothing else calling into the plugin.

Signed-off-by: Eric Blake <eblake@redhat.com>