server: Fix assertion when .can_zero or .can_extents lie
Although it is unlikely in practice to hit a plugin that advertises
.can_FOO without also supporting .FOO, that is no reason for us to
fail an assertion.  We can only assert the existence of callbacks that
we checked when loading the plugin.  Most of our functions had safe
fallbacks, but .zero and .extents did not.  The assertion in .zero
dates back to 876d715, but was latent until .can_zero was introduced.

Fixes: 7ac9883100
Fixes: 4ca66f70a5
Signed-off-by: Eric Blake <eblake@redhat.com>
(cherry picked from commit d5a4190eb51d023ffcde4720f773db553bcdad64)