cc: Let missing .zero fall back to .pwrite
Failing with EOPNOTSUPP instead of EINVAL is more like python and
other language bindings for sane behavior when .has_zero claimed
support but .zero is missing, because then nbdkit will use .pwrite.

Tested by applying this temporary diff, then using nbdsh to drive
calls to .zero; pre-patch, the calls fail and the image is unchanged,
post-patch, nbdkit still warns about the missing callback, but the
image has zeroes written and the client is unaware of the server
warning.

| diff --git i/tests/shebang.c w/tests/shebang.c
| index 50e07f84..525f7337 100755
| --- i/tests/shebang.c
| +++ w/tests/shebang.c
| @@ -41,12 +41,19 @@ my_pwrite (void *handle, const void *buf,
|    return 0;
|  }
|
| +static int
| +my_can_zero (void *handle)
| +{
| +  return 1;
| +}
| +
|  static struct nbdkit_plugin plugin = {
|    .name              = "myplugin",
|    .open              = my_open,
|    .get_size          = my_get_size,
|    .pread             = my_pread,
|    .pwrite            = my_pwrite,
| +  .can_zero = my_can_zero,
|  };
|
|  NBDKIT_REGISTER_PLUGIN(plugin)

Fixes: a31efe33
Signed-off-by: Eric Blake <eblake@redhat.com>