api: Remove .list_exports from nbdkit 1.22 release.
During the 1.21 development cycle we added support for listing
exports.  However at the time that 1.22 was released we did not feel
the API was sufficiently finalized to commit to in a stable branch.
Therefore this new API was removed before 1.22 was released, but the
feature continues to be developed upstream and should appear in
nbdkit 1.24.

Note this also hides the nbdkit*export* functions so they can only be
used from filters (filters do not have a stable API).