ext2: Supply .list_exports and .default_export
When using ext2file=exportname to pick the file to server from the
client's export name, we do not want to leak the underlying plugin's
export list.  While touching this, take advantage of string lifetimes
via nbdkit_strdup_intern and similar for less cleanup bookkeeping.

Note that we don't actually implement a full .list_exports; doing that
with NBD_OPT_LIST is likely prohibitive (it's likely the disk contains
LOTS of files), and would require that we can do next_ops->pread prior
to the client reaching our .open.  For the former issue, it would be
better if the NBD protocol adds a new option NBD_OPT_LIST_HIER that
has replies that differentiate between containers and leaves, and only
lists elements within a container supplied as an argument.  For the
latter, we already know we want to tweak filters to be able to open a
plugin independently of a client's connection.  And for now we punt,
and leave the advertised list empty.

It's also high time we had some testsuite coverage of
ext2file=exportname.  Although ext2.img doesn't technically depend on
the new test, including the dependency ensures the file gets rebuilt
in an incremental build.

Signed-off-by: Eric Blake <eblake@redhat.com>