common/utils: Add make_temporary_directory() function.
It's not possible in Windows to create a replacement mkdtemp()
function.  The reason is that call sites use code like:

  char template[] = "/tmp/XXXXXX";
  mkdtemp (template);

but /tmp/ is not a meaningful location on Windows, and there is not
enough space in the string to place a meaningful path in the template.

Since fixing this would require changing every call site anyway, we
might as well create a new function for creating and returning a
temporary directory.

I only changed one place where we use mkdtemp (in the server), but
there are many other places to change although those are in
plugins/filters which cannot be ported to Windows yet for other
reasons.