nbd: Implement .default_export, .export_description
As long as we can only connect to one export name of the server, we
don't need .list_exports, and our default export name was hard-coded
on the command line.  With new libnbd 1.4 functionality, we can also
report any description that the server handed us.

(Of course, limiting ourselves to one export name is boring, so the
next patch will take further advantage of libnbd 1.4 to lift that
restriction.)

Signed-off-by: Eric Blake <eblake@redhat.com>