nbd: Add dynamic-export=true option
Now that nbdkit gives us the ability to know what export name the
client wants, we can pass that name on to the server.  This only works
for non-shared connections (in a shared connection, we only connect to
the server once, but we can't differentiate content without different
connections); and when uri is specified, it requires libnbd 1.4 (as we
need opt_mode to override the export name from the uri with the export
name from the client).

libnbd 1.4 also lets us report the canonical name of the "" export
according to the server.  With this patch, it will merely change our
advertisement from [""] to ["name"], but the next patch will further
add .list_exports so that a client can see the same advertisement that
the server sent.

Signed-off-by: Eric Blake <eblake@redhat.com>