tests: Don't require daemonizing nbdkit to use start_nbdkit function.
On Unix-like systems tests/functions.sh:start_nbdkit relies on nbdkit
to daemonize itself.

This commit changes the function so that on Windows (only) it runs
“nbdkit -f .. &” (ie. runs nbdkit in the foreground and then
backgrounds it in the shell).  This is so we can run these tests on
Windows which does not yet support daemonizing.

The commit also removes requires_daemonizing from those tests which
are now able to run successfully.