api: Add nbdkit_str[n]dup_intern helper
An upcoming patch wants to introduce .default_export and
.export_description callbacks that each return a string.  We could
have them return 'char *' where the caller creates the result with
malloc() and nbdkit calls free() on the result, but that seems like an
awkward coupling of the ABI, and is wasteful for plugins that can
return a compile-time constant.  So we'd rather have it return 'const
char *'.  But that leads to a different problem: if the string is
generated dynamically, the plugin has to guarantee a lifetime that
lasts beyond the callback return (otherwise nbdkit would commit
use-after-free), but that forces the plugin to track the memory
(otherwise the plugin would have a memory leak), and to eventually
clean it all up during .unload or .close.  Worse, for .default_export,
our intended use is calling it prior to .open, where there is no
available handle for the plugin to track the allocation, short of
tracking it globally for the eventual .unload.  But even though such
memory is tracked, it behaves like a memory leak, in that repeated
client connections cause further allocations that are not cleaned up
in the short term, rather than a steady-state usage of memory per
client.

Avoiding malloc()/free() in the ABI is also nice for the sake of
internal implementation: if all allocation/freeing is done on the same
side of the fence (either both in the plugin, or both in nbdkit), then
we can swap out allocators under the hood, such as playing with a pool
allocator instead of malloc(), to see if we can get performance
improvements; but if the allocator is part of the ABI, then swapping
implementations without breaking binary compatibility becomes a lot
harder, possibly requiring additional data copying.

So, the solution is to let nbdkit track an alternate string on behalf
of the plugin, where nbdkit then cleans it up as the client goes away.

For reference, we already have examples of an interface where we hid
the allocation under the hood: filters use nbdkit_extents_new() so
that all malloc()/free() is done by nbdkit.  We also have API where
the allocator leaks across boundaries: nbdkit_realpath() and
nbdkit_absolute_path() require the plugin to use free().  So a future
patch may want to add nbdkit_realpath_intern() which returns a const
char *, and leaves the cleanup to nbdkit rather than the plugin.

There are several places in the existing code base that can also take
advantage of this copying functionality, including in filters that
want to pass altered strings to .config while still obeying lifetime
rules.

Signed-off-by: Eric Blake <eblake@redhat.com>