test-layers: Rewrite to use libnbd
As mentioned in the test itself, writing our own ad-hoc NBD client is
not easily maintainable.  Rewriting the test to utilize libnbd reduces
the amount of code we have to maintain, while making it possible for
the next patch to add coverage for things that we documented as too
complex with our ad-hoc client.

Signed-off-by: Eric Blake <eblake@redhat.com>