test-layers: Add coverage of .list_exports
Fixes a testing gap that was not possible to test in d4c58f93, back
when we had an ad-hoc client.  This rearranges the test to check for
.load and friends as soon as possible, while waiting to check
.get_size and friends until after we have finished .open.

Note that the test of .list_exports added in the test is only run
under libnbd 1.4; but the test is not skipped even with older libnbd
because the rest of the test is still functionally useful even on the
subset that gets tested.

Signed-off-by: Eric Blake <eblake@redhat.com>