api: Add .default_export
I'm about to add an 'exportname' filter, and in the process, I noticed
a few shortcomings in our API.  Overloading .list_exports in order to
determine a canonical export name at .open time is awkward; the two
uses (answering NBD_OPT_LIST for a full list, vs. remapping a client's
"" into a canonical name during .open) are orthogonal enough to
warrant separate plugin callbacks.  This will also make it easier to
express the notion of no default export (connecting to "" is an error)
at the same time as listing other exports.  Another consideration is
that when tls=1, the choice of export to expose pre-TLS vs. post-TLS
may differ, but without a call to .open yet, our just-added
nbdkit_is_tls() does not fit our preferred lifecycle, so this has to
be a parameter to the new .default_export.  We will alter the
signature of .list_exports soon; in the meantime, the bool
default_only parameter is now ignored.

Adding .default_export support to sh/eval is big enough for a separate
patch.

The ondemand plugin continues to advertise "" (and not "default") in
its list of exports, but now reports a canonical name of "default"
when connecting to it.

The cc plugin gets .default_export now; it will get .list_exports when
we fix that signature.

The tls-fallback filter can now avoid leaking a default export name
through an insecure connection.

In the testsuite, test-layers shows that a connection to export ""
invokes the new callback through the chain of backends; while
test-eval-exports is temporarily weakened until sh support is enhanced
later.