tests: Silence python stack trace during expected failure
test-offset-truncate.sh intentionally provokes out-of-bound reads
under the shell function expect_fail; but does so in a way that causes
nbdsh to die with a python stack dump.  Abrt then picked this up and
tried to report it as an application crash.  Improve the test by
gracefully letting python exit with non-zero status without a crash.

Signed-off-by: Eric Blake <eblake@redhat.com>