server: Fix server debug flags (-D nbdkit.*) on Windows.
On Unix-like systems nbdkit.syms is a linker script so it controls
which symbols are exported from the nbdkit server binary.

On Windows it's a little bit different: nbdkit.def (derived from
nbdkit.syms) is used to build an import library which is linked into
plugins and filters and controls what symbols they can see.  However
for server debug flags what we actually want is for server symbols to
be seen from the server, so nbdkit.def has no effect.  In any case it
wouldn't really matter because our current nbdkit.def does not export
any of the debug flags because Windows export definition files don't
support wildcards.

So for Windows what we have to do to make server debug flags work is
extern and export them with __declspec(dllexport).

Note this only applies to server debug flags.  Plugin and filter debug
flags work fine on Windows already (although I'm unclear exactly why
since they are not externed or exported, but hey, magic).