data: Rewrite format parser so it builds an intermediate parse tree.
The language accepted is broadly the same, although this fixes some
parsing bugs in the previous version.  The tests continue to pass
unmodified.

The advantage is it allows more flexible parsing.  For example
previously these were not accepted but now they work:

  <FILE[2:100]
  "String"[:2]
  ( 1 2 3 4 5 6 ) * 10 [:50]

Additionally it makes it easier to implement optimizations and new
features in future.  For example it would be nice if we could write
'</dev/urandom[:16]' (pick 16 bytes of randomness) but at the moment
that would be an infinite loop.  With a proper parse tree we could
optimize this case.

In future it should be possible to add an optimization pass.  There
are many opportunities: eg. turning lists of bytes and nested *N
expressions into a single string.