data: Optimize BYTE*N and <FILE[N:M]
The first one was optimized in the previous ad hoc parser, and it
saves creating a whole allocator for a single byte so makes sense to
do it here.

The second was not possible at all in the ad hoc parser, but was
implemented as a side effect in the new parser.  The optimization in
this commit makes possible "</dev/urandom[:len]" (get len bytes of
randomness) without going into an infinite loop reading the whole of
/dev/urandom.