windows: Don't use CreateEventA() variant.
Win32 API handles strings in a peculiar way such that every function
has two variants -- *A for "ANSI" (8 bit encoding like the old
codepages) and *W (wide) for "Unicode" (specifically UTF-16LE).  There
is usually a non-postfixed macro which invokes either the *A or *W
variant depending on how the program is built.

I examined all the cases in nbdkit where we used the *A variant
explicitly instead of the non-prefixed version.  In all other cases
except this one we actually wanted the "ANSI" variant because we were
using 8 bit strings, but in this case we can use CreateEvent.  (We do
not explicitly use the *W variants anywhere).