common/allocators: Use 64 bit ints for all counts and sizes.
With the data plugin it is possible to create subpatterns which are
larger than a uint32_t.  This causes the plugin to silently overflow
when doing operations like '*'.  eg. This impossible operation worked
before:

$ ./nbdkit data '
           1 *2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2
             *2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2*2 ' \
           --run 'qemu-img info $nbd'
image: nbd://localhost:10809
file format: raw
virtual size: 1 TiB (1099511627776 bytes)

Simply changing all the functions to take uint64_t is sufficient.  The
data plugin is already passing 64 bit ints for counts, and the memory
plugin uses 32 bit ints which will be extended because the functions
are prototyped.

Thanks: Eric Blake