common/allocators/sparse.c: Optimize blit of sparse/zero pages.
When blitting sparse/zero pages into the allocator, test if they are
zero (expected to be a common case) and free the page if so.

We could actually do even better here by using the extents information
from the other allocator to skip the copying entirely, but the
implementation of that would be much more complex.