data: Adjust parser to use a global table of expressions.
To perform more complex optimizations on the AST we will need to
delete nodes, share nodes, etc.  This is difficult in a non-GC
language without using reference counting, which is awkward in itself.

However a simple way to implement it is to have a global table of
expressions and reference these by their index into the table
(node_id).  We can freely share and unreference nodes during
optimizations.  At the end we clean up by deleting the whole table.