tests: Improve python style in all tests using nbdsh
Using a py linter on a .sh test that invokes nbdsh is not practical,
so I probably missed some style changes.  However, it's pretty
mechanical to fix our testsuite to consistently prefer 'method()' over
'method ()' when it comes to python snippets; likewise, a consistent
pattern of handling expected exceptions is easier to follow.

Signed-off-by: Eric Blake <eblake@redhat.com>