tests: Improve test-file-dir.sh
Using 'nbdsh -u $bad_uri' will cause nbdsh 1.4 to leak an nbd.Error
excpetion all the way out of python, which in turn triggers ABRT with
python3-abrt-addon to potentially flag a program crash.  Also, our
logic to test for expected failure is rather convoluted.  Better is to
avoid -u, and instead catch the expected exception ourselves, so that
we can rely on successful exit as our witness that the export did not
exist.

Signed-off-by: Eric Blake <eblake@redhat.com>