api: Add .export_description
I'm about to add an 'exportname' filter, and in the process, I noticed
a few shortcomings in our API.  Having .default_export makes it easy
to answer NBD_INFO_NAME in response to a client request during
NBD_OPT_GO, but answering NBD_INFO_DESCRIPTION is awkward - there's no
guarantee the export name was given with a description in
.list_exports.  Note, however, that while we map "" into a canonical
name prior to .open (as the plugin is easier to write if it can just
assume the client already passed in the canonical name), we don't need
a description until after .open, alongside .get_size.

Note that in general, we cannot require the description given (or
omitted) in .list_exports to match the description returned by
.export_description.  Whether a client asks for a listing of available
exports is orthogonal to whether a client requests a description on a
single export.

The recent libnbd 1.4 release has added 'nbd_set_full_info()' as the
knob for whether a client will request name and description during
NBD_OPT_GO; adding this functionality in nbdkit makes it easier to
test that API in libnbd.  Also note that 'nbdinfo --list' released
with libnbd 1.4 does NOT query descriptions during NBD_OPT_INFO (it
assumes that the description during NBD_OPT_LIST is sufficient); but
later libnbd does as a fallback for when the list description is
empty, making it trickier to write a test that passes across libnbd
versions.