tests/test-exportname.sh: Fix test.
When “nbdkit ... --run 'nbdinfo ...' > out” is expected to fail, we
cannot be sure of the content of the "out" file.  In particular
nbdinfo produces its output incrementally, so in JSON mode there's
usually a stray "{" in the output, and in non-JSON mode it usually
prints the "protocol: " line.

Thus it's not correct to test for the output file being non-empty.

Remove these lines which caused the test to fail.  However I replaced
them with "cat" for diagnostic purposes.

Fixes: commit 7623b2cc45078cca88fdd2d96c70c7f82a0db49d