server: Support --run and --vsock.
It's possible to support this as long as the captive script uses
"$uri" (not $nbd) and the command supports nbd+vsock URIs.

This also modifies the test to use these options together.  However
that revealed a bug in libxml2 which limits port numbers to 99,999,999
so I had to add a hopefully temporary workaround for that.