tests: In some tests replace qemu-img convert with nbdcopy.
This was mainly done to allow us to give nbdcopy wider testing.  We
can be fairly sure that qemu-img convert works as it is already very
widely used so there's no reason to test qemu-img.

However nbdcopy also has some novel features which we can take
advantage of.  In particular it will usually use multi-conn which
exercises new code paths in nbdkit.  It is also more aggressive about
reading and writing out of order in the default configuration.