tests/test-nbdkit-backend-debug.sh: Fix test using nbdkit noextents filter.
This test relied on the client issuing NBD_CMD_READ [pread] commands.
However the test disk (nbdkit-memory-plugin) is completely sparse so a
client which starts by examining the extents of the disk will never
issue pread commands.  For some reason qemu-img convert did issue
pread, but it's not guaranteed it will stay that way in future.

The simple solution to this is to disable extents by adding the
noextents filter on top.

We can go further by replacing nbdkit-nofilter-filter with noextents
since the only reason to use nofilter in the first place was because
we needed any filter to test message output.