tests: Use nbdcopy for test-rate.sh and test-rate-dynamic.sh.
As in the previous commit, some tests now use nbdcopy instead of
qemu-img convert.

Initially these modified tests of the rate filter took a lot longer to
run when using nbdcopy vs qemu-img convert.  I investigated this and
found that the rate filter is behaving correctly and as documented,
but unexpectedly.

nbdcopy in its default configuration will request the entire 25M of
the test disk in a single pread request.

Because the token bucket abstraction used in the rate filter has a
“capacity” setting (hard-coded to 2.0 * rate-equivalent seconds) when
the request for an entire disk comes in we sleep for the required 18
seconds but only add 2 rate-equivalent seconds worth of tokens to the
bucket, thus needing to sleep again for 16, 14, [etc] seconds
(90 seconds in total).

One way to fix this would be to make the token bucket much larger, but
I chose instead to use nbdkit-blocksize-filter to break up the large
requests from nbdcopy into much smaller ones which does not hit this
peculiar lumpy behaviour of the rate filter.