server: Make $nbd be a simple synonym for $uri.
While I was troubleshooting a problem with an nbdkit user recently,
they had a problem understanding the difference between $nbd and $uri.
In the past it was necessary to construct magical paths for qemu or
guestfish but both these tools now support standard NBD URIs (and if
we discover cases where they do not, we should fix those tools).

This change makes $nbd be a synonym for $uri, removing the magic.

If you are using very old qemu/guestfish you can still construct the
old paths manually using the other variables, but I would advise
upgrading.