log: Introduce enter()/leave() and other helper functions.
The existing code didn't make a very clear distinction between the
message printed on entry and exit from filter functions.  This commit
cleans up the code by introducing enter() and leave() functions for
the two different tasks, and a print() function for the cases which do
not have separate messages.  Note the change here is how the "..." in
the message is handled.

For many functions we further introduce a LOG macro that calls enter()
on entry and leave_simple() on any return path, implemented using
attribute((cleanup)).

This commit should have no effect on the log file output.