tests: Add requires_libnbd_version to test for minimum version of libnbd.
https://github.com/libguestfs/libnbd/commit/1bacff644e949688ba5e7b7ba412ed50efab85ac
changed the output of commands like ‘nbdcopy --version’ from:

  $ nbdcopy --version
  libnbd 1.5.7

to:

  $ nbdcopy --version
  nbdcopy 1.5.9
  libnbd 1.5.9

This broke the existing requires test, so these tests would never have
run.  This is an opportunity to add a requires_libnbd_version function
for tests which does this properly.

Set the minimum version for these tests to 1.5.9 (the current upstream
version).  It could work with earlier versions, but this is the one I
have tested and a lot of fixes went into nbdcopy recently.