server: Print server name and version as first debug message.
Currently the first debug message is this one (assuming TLS is
disabled which it usually is):

  nbdkit: debug: TLS disabled: could not load TLS certificates

What would be more useful as a first debug message, especially when
debugging problems at customer sites, would be the version of nbdkit.
This simple commit changes the first messages to:

  nbdkit: debug: nbdkit 1.23.11
  nbdkit: debug: TLS disabled: could not load TLS certificates

This of course has no effect when debugging is not enabled.