server: Print server name and version as first debug message.
Currently the first debug message is this one (assuming TLS is
disabled which it usually is):

  nbdkit: debug: TLS disabled: could not load TLS certificates

What would be more useful as a first debug message, especially when
debugging problems at customer sites, would be the version of nbdkit.
This simple commit changes the first messages to:

  nbdkit: debug: nbdkit 1.23.11
  nbdkit: debug: TLS disabled: could not load TLS certificates

This of course has no effect when debugging is not enabled.

Note for stable-1.22:

Although this is a new feature, it's immensely useful for debugging
and has already been tried out in RHEL.

(cherry picked from commit 3b84a9d40e1aa5e20cf84a0d18b32a9ecd152e3d)