server: Add ./configure --with-extra=... for downstream packagers.
Similar to libguestfs, allow a way for downstream packagers to add a
freeform string that can be used for troubleshooting rebuilds.

In Fedora’s spec file I intend to add:

  ./configure --with-extra='%{name}-%{version}-%{release}'

which would mean the version is printed as:

  $ nbdkit --version
  nbdkit 1.23.12 (nbdkit-1.23.12-1.fc34)

Note for stable-1.22:

Although this is a new feature it's very useful for downstream
packagers so I have decided to include it in nbdkit 1.22.

(cherry picked from commit 15354d8c64191028bb7edf4a287664cb960b26b2)