streaming, tests: Fix for new, lower disk size limit in qemu > 5.2.0.
This qemu commit reduced the maximum disk size supported from
2⁶³ - 512 bytes to 2⁶³ - 2³⁰ bytes:

  https://git.qemu.org/?p=qemu.git;a=commitdiff;h=8b1170012b1de6649c66ac1887f4df7e312abf3b

Since nbdkit was testing qemu’s older limit, we must adjust the
streaming plugin and a couple of tests which embedded this limit.

Unfortunately this is not the end of the bugs.  I discovered that
qemu-io (and possibly qemu parsing in general) uses floating point so
silently truncates everything at 53 bits.  Any large number which is
not a multiple of 2048 could be silently misinterpreted by the qemu
command line:

  https://www.mail-archive.com/qemu-devel@nongnu.org/msg770572.html

Thanks: Miroslav Rezanina, Daniel Berrangé