tests/old-plugins: Remove libselinux dependency from old plugins.
Thomas Weißschuh reported that it was not possible to test these
plugins on a platform which doesn't have libselinux (Arch in that
case) because they linked with that library.  The reason they linked
with that library was because of an overlinking bug (corrected in
commit 6c51add6bef8a08c99cb3e1d37ce6cb1812dba9d).

I fixed this by backporting this commit to stable-1.2 and recompiling
the plugins for 32- and 64-bit, producing plugins that no longer link
with libselinux.