cow: Implement efficient trimming support.
By storing an extra bit in the overlay bitmap we can mark when whole
blocks have been trimmed.  This allows us to match a feature of qcow2
(zero clusters).  This is much more efficient than what happened
before where trims would be turned into zero writes causing the size
of the overlay to balloon.

In theory we could punch holes in the overlay to save a bit of disk
space, but this patch does not implement this yet.