cache, cow: Do not round size down.
Both the cache and cow filters have the annoying behaviour that the
size of the underlying plugin is rounded down to the block size used
by the filters (ie. 4K).  This is especially visible for single sector
disk images:

$ ./nbdkit memory size=512 --filter=cow --run 'nbdinfo --size $uri'
0
$ ./nbdkit memory size=512 --filter=cache --run 'nbdinfo --size $uri'
0

but affects all users of these filters in a way which appears random
and unexpected.

Getting rid of this assumption is awkward: the general idea is that we
round up the size of the backing file / bitmap to a full block.  But
whenever we are asked to read or write the final block to the
underlying plugin we have to handle the tail case carefully.

This also adds tests for various corner cases.