readahead: Fix multi-conn inconsistency
The readahead filter was blindly passing through the plugin's
.can_multi_conn, but does not kill the readahead buffer on a flush.
This is a bug in the following scenario, when the plugin advertised
multi-conn:

connA              connB
write
  wipes our buffer
  plugin caches the write
                   read
		     reads stale data into buffer
flush
  plugin finally flushes the write
                   read
		     data in buffer is still stale

In order to preserve the plugin's multi-conn status, we MUST drop our
cache any time we flush, so that our next read picks up whatever got
flushed from another connection.

Even when the server lacks multi-conn, we still want to kill our
readhead buffer, as we hope that a flush in connA to write data to
disk, then a flush on connB to reload from disk rather than from
cache, will be enough to let connB see what connA wrote.  Note,
however, that it is not guaranteed (the NBD spec does not require
FLUSH to be consistent across connections if multi-conn was not set).